requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
click==7.1.2
cycler==0.10.0
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.2
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
html5lib==1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.1.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
reportlab==3.5.67
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.0
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.5
